be_booktracker/requirements.txt
Django==4.2.16
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
psycopg2-binary==2.9.10
python-dotenv==1.0.1